+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Added
+* Added an option to generate meshes with _speff materials on export for Xenoblade 3.
+
 ### Changed
 * Improved the layout of nodes in generated materials.
 * Changed the name of materials to include an index prefix to handle materials with duplicate names.